feat: `useFieldGroup` composable feat: added `checkChildFieldGroups` property to `FieldGroup`
