+++ docs/src/pages/api/field-group.mdx
@@ -45,6 +45,13 @@ It can be nested which means there can be `FieldGroup` inside `FieldGroup` (like
 
 ## API Reference
 
+### Props
+### Props
+
+| Prop                  | Type                           | Required/Default | Description                                                                                                                                                                                                                     |
+| :-------------------- | :----------------------------- | :--------------- | :------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| checkChildFieldGroups                    | `boolean`                       | `true`         | Determines if the field group's final state will also check nested field groups |
+
 ### Slots
 
 #### default
